requirements.txt
python-dotenv==1.0.0
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
email-validator==2.1.0
python-dateutil==2.8.2
Werkzeug==3.0.1
gunicorn==21.2.0
waitress==2.1.2
psycopg2-binary==2.9.9